pyproject.toml
[project]
name = "tuntap-demo"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "click>=8.1.8",
    "logging-tree>=1.10",
    "pyserial>=3.5",
    "python3-pytun>=1.0.2.2",
    "rich>=13.9.4",
    "scapy>=2.6.1",
]

[tool.uv]
package = true

[project.scripts]
taptool = "tuntap.main:entrypoint"
